requirements.txt
pandas
matplotlib
python-docx
openpyxl
streamlit
Pillow